[FLINK-13066][hive] append hive-site.xml to path of Hive conf dir
This PR fixes a bug that we previously only pass Hive conf dir to HiveConf but we really should pass path of hive-site.xml. Thus, the change is to append hive-site.xml to the Hive conf dir and pass into HiveConf if Hive conf dir is not null.

This closes #8955.