[FLINK-18537][table] Replace RowDataTypeInfo with InternalTypeInfo
TypeInformation is a legacy class for the sole purpose of creating a
TypeSerializer. Instances of TypeInformation are not required in the
table ecosystem but sometimes enforced by interfaces of other modules
(such as org.apache.flink.api.dag.Transformation). Therefore, we
introduce InternalTypeInfo which acts as an adapter whenever type
information is required. Instances of InternalTypeInfo should only
be created for passing it to interfaces that require type information.
The class should not be used as a replacement for a LogicalType.
Information such as the arity of a row type, field types, field names, etc.
should be derived from the LogicalType directly.

This closes #12900.